+++ tests/conftest.py
@@ -3,7 +3,7 @@ import asyncio
 import pytest
 
 @pytest.fixture(scope='module', autouse=True)
-def clear_files_teardown():
+def _teardown():
     yield None
     # prevent ResourceWarning (in devel mode)
     asyncio.get_event_loop().close()